refactor(html): extract the Mux element behavior into a mixin
The `poster-time` attribute, `src` reflection, and the storyboard `<track>`
child are all Mux identity, not engine behavior, but they lived on a class
built over the hls.js-backed Media. A second Mux flavor would need the same
behavior over a different `CustomMediaElement`, and there is no common class
to extend — only a common host contract, so a mixin is what fits.

Drops the `@videojs/media/dom/hls-js` dependency along the way: the live
check now reads `MediaStreamTypes` from the engine-neutral surface.
